Create nvda.po
Hi. The NVDA translation system refuses to accept this translation. When I send it, it is sent, but when I try to update it (e.g. to fix something) and send the updated one, an error occurs asking me to update the working copy. When I agree, it returns an old file that contains only seven lines. Please help me fix this problem.