Updated Ukrainian translation
Please include this translation in the addon. I have updated not only the translation of the addon, but also the translation of the documentation. I cannot send the translation via the NVDA translation system because the file in the Ukrainian (UK) folder is corrupted (it contains only eight lines). I also sent you an updated translation last year, but you ignored it. Please accept the pull request this time and include it in the addon! Thank you!